refactor(logging): create LogLevelBadge component for log level display
- Removed inline log level color definitions from LogCard.
- Introduced a new LogLevelBadge component to handle log level styling and rendering.
