+++ v1-0-0/src/components/logging/log-card.tsx
@@ -2,19 +2,9 @@
 
 import { ChevronDown, ChevronRight } from "lucide-react";
 import { Badge } from "@/components/ui/badge";
-import { LogEntry, LogLevel } from "@/hooks/use-logger";
+import { LogEntry } from "@/hooks/use-logger";
 import { formatDate } from "@/lib/date-utils";
-
-const LOG_LEVEL_COLORS: Record<LogLevel, string> = {
-  error:
-    "bg-red-500/10 text-red-700 border-red-200 dark:text-red-400 dark:border-red-800",
-  warn: "bg-yellow-500/10 text-yellow-700 border-yellow-200 dark:text-yellow-400 dark:border-yellow-800",
-  info: "bg-blue-500/10 text-blue-700 border-blue-200 dark:text-blue-400 dark:border-blue-800",
-  debug:
-    "bg-purple-500/10 text-purple-700 border-purple-200 dark:text-purple-400 dark:border-purple-800",
-  trace:
-    "bg-gray-500/10 text-gray-700 border-gray-200 dark:text-gray-400 dark:border-gray-800",
-};
+import { LogLevelBadge } from "./log-level-badge";
 
 interface LogCardProps {
   entry: LogEntry;
@@ -50,12 +40,7 @@ export function LogCard({ entry, isExpanded, onToggleExpand }: LogCardProps) {
                 {`[${formatDate(entry.timestamp)}]`}
               </span>
 
-              <Badge
-                variant="outline"
-                className={`${LOG_LEVEL_COLORS[entry.level]}`}
-              >
-                {entry.level.toUpperCase()}
-              </Badge>
+              <LogLevelBadge level={entry.level} />
 
               <Badge variant="secondary">{entry.context}</Badge>
